test(send): refactor useSendForm hook tests
